requirements.txt
flask==2.0.3
werkzeug==2.0.3
flask-wtf==1.0.1
google-api-python-client==2.97.0
python-dotenv==1.0.0
requests==2.31.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
python-dateutil==2.8.2 